[flights] moved logic to a separate util file
