+++ ComfortZoneApp/app/src/main/java/com/example/comfortzone/Fragments/FlightFragment.java
@@ -1,7 +1,8 @@
 package com.example.comfortzone.fragments;
 
+import static com.example.comfortzone.utils.WeatherDbUtil.maybeUpdateCitiesList;
+
 import android.os.Bundle;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -13,24 +14,16 @@ import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.example.comfortzone.AllWeathersDatabase;
-import com.example.comfortzone.CityListCallback;
 import com.example.comfortzone.FlightsAdapter;
-import com.example.comfortzone.GroupUrlCallback;
 import com.example.comfortzone.R;
-import com.example.comfortzone.WeatherClient;
-import com.example.comfortzone.GetWeatherCallback;
 import com.example.comfortzone.models.WeatherData;
-import com.example.comfortzone.models.WeatherGroupData;
-import com.google.gson.Gson;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.concurrent.TimeUnit;
 
 public class FlightFragment extends Fragment {
 
     public static final String TAG = "FlightFragment";
-    private WeatherClient client;
     private FlightsAdapter flightsAdapter;
     private List<WeatherData> cityList;
     private RecyclerView rvCities;
@@ -51,13 +44,12 @@ public class FlightFragment extends Fragment {
     public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
 
-        client = new WeatherClient();
         cityList = new ArrayList<>();
         flightsAdapter = new FlightsAdapter(getContext(), cityList);
         db = AllWeathersDatabase.getDbInstance(getContext().getApplicationContext());
 
         initViews(view);
-        maybeUpdateCitiesList();
+        maybeUpdateCitiesList(getContext());
         populateViews();
     }
 
@@ -73,50 +65,8 @@ public class FlightFragment extends Fragment {
         flightsAdapter.addAll(db.weatherDao().getAll());
     }
 
-    public void maybeUpdateCitiesList() {
-        if (db.weatherDao().getAll().isEmpty()) {
-            client.getCityData(new CityListCallback() {
-                @Override
-                public void onGetCityList(WeatherData[] weatherData) {
-                    db.weatherDao().insertAll(weatherData);
-                }
-            });
-        }
-        Long timeNow = System.currentTimeMillis();
-        Long hourAgo = timeNow - TimeUnit.HOURS.toMillis(1L);
-        if (db.weatherDao().isUploadedOverHrAgo(hourAgo)) {
-            getAndStoreWeatherData();
-        }
-    }
 
-    public void getAndStoreWeatherData() {
-        client.getGroupWeatherUrl(getContext(), new GroupUrlCallback() {
-            @Override
-            public void onGetWeatherUrlGroupIds(List<String> groupUrls) {
-                for (String apiUrl : groupUrls) {
-                    client.getGroupWeatherData(apiUrl, new GetWeatherCallback() {
-                        @Override
-                        public void onGetWeatherData(String data) {
-                            Gson gson = new Gson();
-                            WeatherGroupData weathers = gson.fromJson(data, WeatherGroupData.class);
-                            saveCities(weathers);
-                        }
-                    });
-                }
-            }
-        });
-    }
 
-    public void saveCities(WeatherGroupData weathers) {
-        WeatherData[] weatherData = weathers.getWeathers();
-        AllWeathersDatabase db = AllWeathersDatabase.getDbInstance(getContext().getApplicationContext());
-        for (WeatherData weather : weatherData) {
-            int id = weather.getId();
-            WeatherData weatherFromDb = db.weatherDao().getWeatherById(id);
-            weatherFromDb.setTempData(weather.getTempData());
-            weatherFromDb.setTimeUploaded(weather.getTimeUploaded());
-            db.weatherDao().updateWeatherData(weatherFromDb);
-        }
-    }
+
 
 }
